despliegue_GCP/requirements.txt
requests==2.31.0
google-cloud-storage==2.11.0
